% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_if_.R
\name{mutate_if_cat}
\alias{mutate_if_cat}
\alias{mutate_if_cont}
\title{mutate_if_cat, mutate_if_cont}
\usage{
mutate_if_cat(.tbl, .funs, ...)

mutate_if_cont(.tbl, .funs, ...)
}
\arguments{
\item{.tbl}{table}

\item{.funs}{}
}
\value{
An object of class \code{.tbl} in with columns mutated according to whe
}
\description{
mutates only types of objects
}
\details{
These are just wrappers around \code{mutate_if} where the \code{.predicate} is \code{is_cat}
and \code{is_cont}.
}
\examples{

  data(iris)
  class(iris$Species)
  iris2 <- iris \%>\% mutate_if_cat( as.character )
  class(iris2$Species)  # character

  iris \%>\% mutate_if_cont( add, 2 )

}
\seealso{
\code{\link[dplyr:mutate_if]{dplyr::mutate_if()}}
}
