% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cat.R
\name{select_cat}
\alias{select_cat}
\alias{select_cat.default}
\alias{select_cat.data.table}
\alias{select_cont}
\alias{select_cont.default}
\alias{select_cont.data.table}
\title{select_cat, select_cont}
\usage{
select_cat(.data)

\method{select_cat}{default}(.data)

\method{select_cat}{data.table}(.data)

select_cont(.data)

\method{select_cont}{default}(.data)

\method{select_cont}{data.table}(.data)
}
\arguments{
\item{.data}{table}
}
\value{
Returns a table-like object of the same class as data unless there are no
columns in which case `NULL` is returned
}
\description{
Select columns by type
}
\details{
select_cat() and select_cont() return only the categorical and continuous
types respectively. This is closely mirrors the dplyr function select but
works with non-table values as well.
}
\examples{

  data(iris)
  select_cat(iris)
  select_cont(iris)

\dontrun{
  setDT(iris)
  select_cat(iris)
  select_cont(iris)
}
}
